feat: use SmallVec<[u32; 16]> for NodePath to reduce allocations
Replace Vec<usize> with SmallVec<[u32; 16]> for DOM path storage.
This avoids heap allocations for paths up to 16 levels deep (covers
virtually all real-world DOMs). Using u32 instead of usize halves
memory per element on 64-bit systems.

- Add smallvec dependency to hotmeal
- Enable smallvec feature in facet
- Update compute_path to return SmallVec<[u32; 16]>
- Update navigate_slot_path/get_slot_parent to take &[u32]
- Add fuzz corpus from recent fuzzing runs
- Gitignore profile.json.gz

Closes #10

+++ hotmeal/src/dom.rs
@@ -247,22 +247,24 @@ impl Document {
     /// Path `[slot, a, b, c]` means: get slot root, then navigate a → b → c.
     fn navigate_slot_path(
         &self,
-        path: &[usize],
+        path: &[u32],
         slots: &HashMap<u32, NodeId>,
     ) -> Result<NodeId, DiffError> {
         if path.is_empty() {
             return Err(DiffError::EmptyPath);
         }
 
-        let slot = path[0] as u32;
+        let slot = path[0];
         let slot_root = *slots.get(&slot).ok_or(DiffError::SlotNotFound { slot })?;
 
         let mut current = slot_root;
         for &idx in &path[1..] {
             let mut children = current.children(&self.arena);
             current = children
-                .nth(idx)
-                .ok_or(DiffError::PathOutOfBounds { index: idx })?;
+                .nth(idx as usize)
+                .ok_or(DiffError::PathOutOfBounds {
+                    index: idx as usize,
+                })?;
         }
 
         Ok(current)
@@ -272,14 +274,14 @@ impl Document {
     /// Path `[slot, a, b, c]` returns (node at [slot, a, b], position c).
     fn get_slot_parent(
         &self,
-        path: &[usize],
+        path: &[u32],
         slots: &HashMap<u32, NodeId>,
     ) -> Result<(NodeId, usize), DiffError> {
         if path.len() < 2 {
             return Err(DiffError::EmptyPath);
         }
 
-        let position = path[path.len() - 1];
+        let position = path[path.len() - 1] as usize;
         let parent_path = &path[..path.len() - 1];
         let parent_id = self.navigate_slot_path(parent_path, slots)?;
